Split ProcessBuilder into Impl and actual class
